+++ js/simpletextbox/Index.svelte
@@ -1,42 +1,19 @@
 <svelte:options accessors={true} />
 
 <script lang="ts">
-	import type { Gradio } from "@gradio/utils";
+	import type { SimpleTextboxProps, SimpleTextboxEvents } from "./types";
+	import { Gradio } from "@gradio/utils";
 	import { BlockTitle } from "@gradio/atoms";
 	import { Block } from "@gradio/atoms";
 	import { StatusTracker } from "@gradio/statustracker";
-	import type { LoadingStatus } from "@gradio/statustracker";
 	import { tick } from "svelte";
 
-	export let gradio: Gradio<{
-		change: never;
-		submit: never;
-		input: never;
-		clear_status: LoadingStatus;
-	}>;
-	export let label = "Textbox";
-	export let elem_id = "";
-	export let elem_classes: string[] = [];
-	export let visible: boolean | "hidden" = true;
-	export let value = "";
-	export let placeholder = "";
-	export let show_label: boolean;
-	export let scale: number | null = null;
-	export let min_width: number | undefined = undefined;
-	export let loading_status: LoadingStatus | undefined = undefined;
-	export let value_is_output = false;
-	export let interactive: boolean;
-	export let rtl = false;
+	const props = $props();
+	const gradio = new Gradio<SimpleTextboxEvents, SimpleTextboxProps>(props);
 
 	let el: HTMLTextAreaElement | HTMLInputElement;
 	const container = true;
-
-	function handle_change(): void {
-		gradio.dispatch("change");
-		if (!value_is_output) {
-			gradio.dispatch("input");
-		}
-	}
+	let old_value = $state(gradio.props.value);
 
 	async function handle_keypress(e: KeyboardEvent): Promise<void> {
 		await tick();
@@ -46,43 +23,48 @@
 		}
 	}
 
-	$: if (value === null) value = "";
-
-	// When the value changes, dispatch the change event via handle_change()
-	// See the docs for an explanation: https://svelte.dev/docs/svelte-components#script-3-$-marks-a-statement-as-reactive
-	$: value, handle_change();
+	$effect(() => {
+		if (old_value != gradio.props.value) {
+			old_value = gradio.props.value;
+			gradio.dispatch("change");
+		}
+	});
 </script>
 
 <Block
-	{visible}
-	{elem_id}
-	{elem_classes}
-	{scale}
-	{min_width}
+	visible={gradio.shared.visible}
+	elem_id={gradio.shared.elem_id}
+	elem_classes={gradio.shared.elem_classes}
+	scale={gradio.shared.scale}
+	min_width={gradio.shared.min_width}
 	allow_overflow={false}
 	padding={true}
 >
-	{#if loading_status}
+	{#if gradio.shared.loading_status}
 		<StatusTracker
-			autoscroll={gradio.autoscroll}
+			autoscroll={gradio.shared.autoscroll}
 			i18n={gradio.i18n}
-			{...loading_status}
-			on:clear_status={() => gradio.dispatch("clear_status", loading_status)}
+			{...gradio.shared.loading_status}
+			on_clear_status={() =>
+				gradio.dispatch("clear_status", gradio.shared.loading_status)}
 		/>
 	{/if}
 
 	<label class:container>
-		<BlockTitle {show_label} info={undefined}>{label}</BlockTitle>
+		<BlockTitle show_label={gradio.shared.show_label} info={undefined}
+			>{gradio.shared.label}</BlockTitle
+		>
 
 		<input
 			data-testid="textbox"
 			type="text"
 			class="scroll-hide"
-			bind:value
+			bind:value={gradio.props.value}
 			bind:this={el}
-			{placeholder}
-			disabled={!interactive}
-			dir={rtl ? "rtl" : "ltr"}
+			placeholder={gradio.props.placeholder}
+			disabled={!gradio.shared.interactive}
+			dir={gradio.props.rtl ? "rtl" : "ltr"}
+			on:input={() => gradio.dispatch("input")}
 			on:keypress={handle_keypress}
 		/>
 	</label>